Secures pages (not solved)

+++ src/com/webappcafe/servlet/product/DeleteCustomer.java
@@ -1,14 +1,16 @@
 package com.webappcafe.servlet.product;
 
-	
+
 import com.webappcafe.dao.CustomerDAO;
 import com.webappcafe.dao.CustomerDAOImpl;
-import java.io.IOException;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.io.IOException;
 
 
 @WebServlet(name = "deleteCustomer", value = {"/deleteCustomer"})
@@ -31,15 +33,24 @@ public class DeleteCustomer extends HttpServlet
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException 
     {
-        String id = request.getParameter("id");
-        CustomerDAO customerDAO;
 
-        if(id != null) 
-        {
-            customerDAO = new CustomerDAOImpl();
-            customerDAO.deleteCustomerByID(Long.parseLong(id));
+        HttpSession session = request.getSession();
+        String username = String.valueOf(session.getAttribute("username"));
+        String password = String.valueOf(session.getAttribute("password"));
+//        Customer customer = (Customer) session.getAttribute("admin");
+
+        if (!username.equals("root") && !password.equals("root")) {
+            response.sendRedirect("landingPage.html");
+        } else {
+            String id = request.getParameter("id");
+            CustomerDAO customerDAO;
+
+            if(id != null)
+            {
+                customerDAO = new CustomerDAOImpl();
+                customerDAO.deleteCustomerByID(Long.parseLong(id));
+            }
         }
-
     }
 
 }
